fix: configure different scenes properly

+++ src/webgpu/shaders/entry.wgsl
@@ -16,7 +16,7 @@ fn vs_main(@builtin(vertex_index) VertexIndex : u32) -> @builtin(position) vec4<
 
 @fragment
 fn fs_main(@builtin(position) pos: vec4<f32>) -> @location(0) vec4<f32> {
-  let seed = u32(uRenderSettings.random_seed);
+  let seed = uRenderSettings.random_seed;
   let pixel = vec2<u32>(u32(pos.x), u32(pos.y));
   let resolution = vec2<u32>(u32(uCanvas.size.x), u32(uCanvas.size.y));
   let frame = u32(uFrameTime.frame);
